requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.35.0
azure-identity==1.23.0
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
ecdsa==0.14.1
fastapi==0.116.0
h11==0.16.0
idna==2.10
iniconfig==2.1.0
msal==1.32.3
msal-extensions==1.3.1
octokitpy==0.15.0
octokitpy-routes==2.5.0
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.4.1
python-jose==3.2.0
requests==2.25.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.35.0
wrapt==1.17.2
